Revert "Feat: Adding better Routing System (#1)"
This reverts commit 267a4be1d280f479de7affeda8fe47466c1b77c3.
