+++ README.md
@@ -764,205 +764,6 @@ And your response should be:
 {"message": "Not Found"}
 ```
 
-## Handling Dynamic URL's
-
-One important aspect of web development is handling dynamic URLs, and what I want to say is that it's normal to have these kinds of URLs.
-```bash
-/users/{id}
-```
-
-And then, we send a request like:
-```bash 
-GET /users/123456
-```
-
-And we need to be able to handle this ID as a parameter, for example:
-```python
-    async def process_dynamic_url(request: Request, name: str):
-        try:
-            return Response(
-                200,
-                {
-                    'msg': f'Hello {name}'
-                }
-            )
-        except Exception as e:
-            print(e)
-
-    server.add_route("/hello_world/{name}", {
-        "GET": process_dynamic_url
-    })
-```
-
-And you might start to think, 'This is so complex; there must be some sophisticated business logic to implement it.' And voilà, this can be achieved by using regex in our `Router` class.
-First, let's understand how regex works and how this can be achieved. We have two strings `"/users/{id}/"` and `"/users/12345/"`. Let's create a prototype:
-```python
-import re
-match = re.match(r'/users/(.+?)/', "/users/123/")
-match.groups()
-"""
-Should return:
-('123',)
-"""
-```
-A simple dive into `r'/users/(.+?)/'`:
-- `()` Captures and stores whatever matches the pattern inside the parentheses.
-- `.` Matches any character except a newline
-- `+` A quantifier that matches one or more of the preceding token
-- `?` Matches as few characters as possible (non-greedy match)."
-
-But there's a better way to handle this regex. We have a feature called 'groups' that allows us to use the `groupdict()` method to get a dictionary. 
-```json
-{
-  "id": "123"
-}
-```
-And this helps us a lot! With this dictionary, we can deconstruct it using `**` and pass it as keyword arguments into our methods, routers, or controllers. 
-Our new regex pattern will be (?P<id>.+?). Let's take a brief dive into this regex!
-- `(?P<id>...)` Captures the matched content as id in our dictionary, just as we want.
-- `.+?` Matches one or more characters, except for a newline, using a non-greedy approach.
-
-```python
-import re
-
-match = re.match(r'^/users/(?P<id>[^/]+?)/$', "/users/123/")
-match.groupdict()
-"""
-Our response should look like this:
-{'id': '123'}
-"""
-```
-
-And I know, you're probably thinking 'Oh god' after seeing this regex, but I'll try to make it as simple as I can!
-
-`^/users/(?P<id>[^/]+?)/$`
-- `^` Indicates the beginning of the string, meaning no characters can appear before `/users`
-- `$` Indicates the end of the string, meaning no characters can appear after `/(?P<id>[^/]+?)/`
-- `(?P<id>...)` Matches anything within this part of the regex and captures it as the group id, which we can reference as `<id>`.
-- `[^/]+?` Matches one or more characters except for newlines and the slash `(/)`, in a non-greedy manner. 
-
-So now we get it! We need to be able to pass something like `/users/{id}` and have it pre-compiled as the regex `^/users/(?P<id>[^/]+?)/$`.
-
-Now, let's return to our Router class. We have an add method that will be responsible for receiving the URL and determining which method should be executed.
-
-```python
-    def add(self, path, methods_handler):
-        self.mapping[path] = methods_handler
-```
-
-Now, our dictionary needs to have the converted URL, not the original one. Let's start by creating a method that will convert these URLs for us.
-```python
-    def parse_dynamic_url(self, url):
-        leading_slash = url.startswith('/')
-        trailing_slash = url.endswith('/')
-        url = url.strip('/')
-
-        pattern = re.compile(r"{(.*?)}")
-
-        url_path = [
-            pattern.sub(lambda m: f"(?P<{m.group(1)}>[^/]+?)", part)
-            for part in url.split('/') if part
-        ]
-
-        path = '/' + '/'.join(url_path) if leading_slash else '/'.join(url_path)
-        if trailing_slash and not path.endswith('/'):
-            path += '/'
-
-        return f"^{path}$"
-```
-
-Let's start with `pattern = re.compile(r"{(.*?)}")`. This code compiles a regex pattern to match any dynamic parts of the URL that should be variable.
-
-For example, with URLs like `/users/{id}` or `/book/{name}`, we need to recognize that there is dynamic input between the `{}` brackets, and this regex helps us identify it.
-
-```python
-url_path = [
-    pattern.sub(lambda m: f"(?P<{m.group(1)}>[^/]+?)", part)
-    for part in url.split('/') if part
-]
-```
-
-We'll split our URL into pieces by the slash (`/`). If any part of the URL matches our regex, it will be converted into our boilerplate format.
-
-For example, `/book/{name}/action/{author}` would be transformed into `/book/(?P<name>[^/]+?)/action/(?P<author>[^/]+?)`.
-
-Now, for the last part:
-```python
-def parse_dynamic_url(self, url):
-        leading_slash = url.startswith('/')
-        trailing_slash = url.endswith('/')
-        ...
-        path = '/' + '/'.join(url_path) if leading_slash else '/'.join(url_path)
-        if trailing_slash and not path.endswith('/'):
-            path += '/'
-
-        return f"^{path}$"
-```
-
-It basically adds a slash at the beginning or the end, if it was originally there. 
-- `/book/{name}/action/{author}` -> `^/book/(?P<name>[^/]+?)/action/(?P<author>[^/]+?)$`
-- `/book/{name}/action/{author}/` -> `^/book/(?P<name>[^/]+?)/action/(?P<author>[^/]+?)/$`
-- `book/{name}/action/{author}/` -> `^book/(?P<name>[^/]+?)/action/(?P<author>[^/]+?)/$`
-
-I highly suggest you debug and test this method on your own, it will make more sense when you see it working.
-
-Now that we have a method to convert our path into a pre-compiled regex, we need to refactor the `add` method.
-```python
-    def add(self, path, methods_handler):
-        self.mapping[self.parse_dynamic_url(path)] = methods_handler
-```
-
-In our `dispatch` method, we need to handle the request by iterating over our mapping array of endpoints and matching each one with our regex. 
-If there's no match, we move to the next endpoint until we find a match or return a `404 Not Found` to the client. 
-```python
-    async def dispatch(self, request: Request):
-        # Iterate through the compiled patterns
-        for url, pattern in self.mapping.items():
-            match = re.match(url, request.url)
-            if not match:
-                continue
-    
-            params = match.groupdict()
-            handler = self.mapping[url].get(request.method)
-    
-            if handler:
-                await self.request_callback_handler(handler, request, **params)
-                return
-        # Handle 404 Not Found if no matching route or method is found
-        await self.response_writer(
-            Response(
-                404,
-                {'message': 'Not Found'}
-            )
-        )
-```
-Once a match is found with our regex, we can use `match.groupdict()` to get a dictionary with our captured groups. Then, we pass it as keyword arguments (`kwargs`) into our route.
-
-And that's it! Now, we just need to add a new route to our `hello_world` file `(python_web_framework/src/routes/hello_world.py)`.
-
-```python
-    async def process_dynamic_url(request: Request, name: str):
-        try:
-            return Response(
-                200,
-                {
-                    'msg': f'Hello {name}'
-                }
-            )
-        except Exception as e:
-            print(e)
-    
-    server.add_route("/hello_world/{name}", {
-        "GET": process_dynamic_url
-    })
-```
-
-With our new route created, we can test it in our terminal with:
-```bash
-curl -X GET http://127.0.0.1:8080/hello_world/Gabs
-{"msg": "Hello Gabs"}
-```
-
 And that's it! Now you have your own simple Python framework. I hope this guide helped you understand how this type of framework works under the hood.
