+++ python_web_framework/src/response.py
@@ -39,4 +39,3 @@ class Response:
             messages.append("\n" + json.dumps(self.body))
 
         return "\n".join(messages)
-
