+++ python_web_framework/src/routes.py
@@ -1,6 +1,6 @@
-from python_web_framework.src.response import Response
 from python_web_framework.src.request import Request
-from python_web_framework.src.server.main import Server
+from python_web_framework.src.response import Response
+from python_web_framework.src.server import Server
 
 
 def hello_world_app(server: Server):
@@ -23,22 +23,7 @@ def hello_world_app(server: Server):
             }
         )
 
-    async def process_dynamic_url(request: Request, name: str):
-        try:
-            return Response(
-                200,
-                {
-                    'msg': f'Hello {name}'
-                }
-            )
-        except Exception as e:
-            print(e)
-
     server.add_route("/hello_world", {
         "GET": hello_world,
         "POST": process_request
-    })
-
-    server.add_route("/hello_world/{name}", {
-        "GET": process_dynamic_url
     })
