+++ tests/client.py
@@ -16,20 +16,6 @@ async def send_get():
     writer.close()
     await writer.wait_closed()
 
-async def send_get_dynamic():
-    message = 'GET /hello_world/gabs HTTP/1.1\r\nHost: example.com\r\n\r\n'
-    reader, writer = await asyncio.open_connection('127.0.0.1', 8080)
-
-    print(f'Send: {message}')
-    writer.write(message.encode())
-
-    data = await reader.read(900)
-    print(f'Received: {data.decode()}')
-
-    print('Close the connection')
-    writer.close()
-    await writer.wait_closed()
-
 
 async def send_post():
     message_template = (
@@ -65,4 +51,4 @@ async def send_post():
 
 
 if __name__ == "__main__":
-    asyncio.run(send_get_dynamic())
+    asyncio.run(send_post())
